AddNutrientsToSelectetRecipes

+++ DelikatessenDrehbuch/Models/Quantity.cs
@@ -3,7 +3,7 @@
     public class Quantity
     {
         public int Id { get; set; }
-        public float Quantitys { get; set; }
+        public double Quantitys { get; set; }
 
 
     }
